Drop Down Mass Selection + Misc. Bug Fixes
Main change with this commit is the addition of the drop down selection
to the precomputed version of the ascot tool.  This removes the inputs
for mass ranges.

Also modified the scale of ranges by a large constant factor.

Also fixes two small bugs in both versions of the tool:
1. Fixed an edge case where a query could be submitted without a
username.
2. Fixed an extraneous error message appearing for now reason.